# docs/site/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Memobase API",
        "summary": "APIs for Memobase, a user memory system for LLM Apps",
        "version": "0.0.21"
    },
    "paths": {
...
        "/api/v1/project/profile_config": {
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "Get Project Profile Config String",
                "operationId": "get_project_profile_config_string_api_v1_project_profile_config_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileConfigDataResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "project"
                ],
                "summary": "Update Project Profile Config",
                "operationId": "update_project_profile_config_api_v1_project_profile_config_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileConfigData",
                                "description": "The profile config to update"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "description": "Create a new user with additional data",
                "operationId": "create_user_api_v1_users_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserData",
                                "description": "User data for creating a new user"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_users__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to retrieve",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update User",
                "operationId": "update_user_api_v1_users__user_id__put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to update",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to update"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "Updated user data",
                                "title": "User Data"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete User",
                "operationId": "delete_user_api_v1_users__user_id__delete",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to delete",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/blobs/insert/{user_id}": {
            "post": {
                "tags": [
                    "blob"
                ],
                "summary": "Insert Blob",
                "operationId": "insert_blob_api_v1_blobs_insert__user_id__post",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to insert the blob for",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to insert the blob for"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlobData",
                                "description": "The blob data to insert"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BaseResponse": {
                "properties": {
                    "data": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data",
                        "description": "Response data payload"
                    },
                    "errno": {
                        "$ref": "#/components/schemas/CODE",
                        "description": "Error code, 0 means success",
                        "default": 0
                    },
                    "errmsg": {
                        "type": "string",
                        "title": "Errmsg",
                        "description": "Error message, empty when success",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "BaseResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProfileConfigData": {
                "properties": {
                    "profile_config": {
                        "type": "string",
                        "title": "Profile Config",
                        "description": "Profile config string"
                    }
                },
                "type": "object",
                "required": [
                    "profile_config"
                ],
                "title": "ProfileConfigData"
            },
            "ProfileConfigDataResponse": {
                "properties": {
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProfileConfigData"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Response containing profile config data"
                    },
                    "errno": {
                        "$ref": "#/components/schemas/CODE",
                        "description": "Error code, 0 means success",
                        "default": 0
                    },
                    "errmsg": {
                        "type": "string",
                        "title": "Errmsg",
                        "description": "Error message, empty when success",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "ProfileConfigDataResponse"
            },
...
        }
    }
}